pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codai"
version = "0.0.8"
authors = [
  { name="Cody", email="cody@dkdc.dev" },
]
description = "AI coding assistant"
readme = "readme.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'python-dotenv',
  'typer[all]',
  'requests',
  'openai',
  'langchain',
  'ibis-framework[duckdb,polars,deltalake]',
  'plotly',
  'streamlit',
]

[project.urls]
"Homepage" = "https://github.com/lostmygithubaccount/codai"
"Bug Tracker" = "https://github.com/lostmygithubaccount/codai/issues"

[project.scripts]
codai= "src.main:app"

